Upgrade the json gem
The previous version of 1.8.3 not able to run in ruby2.5 because of the
merger of the Fixnum and Bignum types.

I am not upgrading all the other gems for the moment because there
appears to be a breaking change somewhere in bourbon/neat/bitters which
makes our stylesheets no longer valid. I plan to investigate that
problem in the future.