Move the CNAME file into source/
This should ensure that the file is always copied into the build
directory. This will ensure the CNAME is correctly located whether the
site is deployed by Github Actions or by a manual deploy.

I have removed the copy command from the Github workflow, because it is
no longer needed.